Rimozione parti inutili
rimozione e sistemazione delle parti critiche del programma
